refactor: centralize property options
