+++ ServerDriveUi/src/main/java/com/example/serverdriveui/ui/component/properties/KeyboardOptionsProperty.kt
@@ -6,6 +6,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import com.example.serverdriveui.service.model.PropertyModel
 import com.example.serverdriveui.ui.state.ComponentStateManager
 import com.example.serverdriveui.util.JsonUtil.asString
+import com.vini.designsystemsdui.PropertyOptions.KeyboardOptionsOption
 
 class KeyboardOptionsProperty(
     private val properties: Map<String, PropertyModel>,
@@ -23,13 +24,6 @@ class KeyboardOptionsProperty(
     override fun getKeyboardOptions() = getValue().toOption().keyboardOptions
 
     override fun setKeyboardOptions(keyboardOptions: KeyboardOptionsOption) = setValue(keyboardOptions.id)
-}
-
-enum class KeyboardOptionsOption(val id: String, val keyboardOptions: KeyboardOptions) {
-    Default("Default", KeyboardOptions.Default),
-    Number("Number", KeyboardOptions(keyboardType = KeyboardType.Number)),
-    Phone("Phone", KeyboardOptions(keyboardType = KeyboardType.Phone)),
-    Password("Password", KeyboardOptions(keyboardType = KeyboardType.Password)),
 }
 
 private fun String?.toOption() =
