+++ ServerDriveUi/src/main/java/com/example/serverdriveui/ui/component/properties/VisualTransformationProperty.kt
@@ -4,10 +4,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import com.example.serverdriveui.service.model.PropertyModel
-import com.example.serverdriveui.ui.component.properties.VisualTransformationOption.None
 import com.example.serverdriveui.ui.state.ComponentStateManager
 import com.example.serverdriveui.util.JsonUtil.asString
 import com.vini.designsystem.compose.textfield.MaskVisualTransformation
+import com.vini.designsystemsdui.PropertyOptions.VisualTransformationOption
 
 class VisualTransformationProperty(
     private val properties: Map<String, PropertyModel>,
@@ -18,7 +18,7 @@ class VisualTransformationProperty(
         properties = properties,
         propertyName = "visualTransformation",
         transformToData = { it?.asString() },
-        defaultPropertyValue = None.id,
+        defaultPropertyValue = VisualTransformationOption.None.id,
     ) {
 
     @Composable
@@ -30,29 +30,7 @@ class VisualTransformationProperty(
 
 private fun String?.toOption() =
     VisualTransformationOption.entries.firstOrNull { it.id == this }
-        ?: None
-
-enum class VisualTransformationOption(
-    val id: String,
-    val visualTransformation: VisualTransformation,
-) {
-    None("None", VisualTransformation.None),
-    Password("Password", PasswordVisualTransformation()),
-    Phone(
-        "Phone",
-        MaskVisualTransformation(
-            mask = "## #####-####",
-            toIgnore = '#'
-        )
-    ),
-    Document(
-        "Documento.CPF",
-        MaskVisualTransformation(
-            mask = "###.###.###-##",
-            toIgnore = '#'
-        )
-    ),
-}
+        ?: VisualTransformationOption.None
 
 interface VisualTransformationComponentProperty {
     @Composable
